chore(fossid-webapp): Move test utility functions to a `TestUtils` file
These functions are going to be used by the Snippet Choice test in a future
commit. Therefore, move them to a `TestUtils` file and change their
visibility to `internal`.
It is also necessary to rename two properties in `FossIdUrlProviderTest`
and remove two properties in `FossIdConfigTest` to avoid clashes.
